pyproject.toml
[project]
name = "relyapi"
version = "0.1.0"
description = ""
authors = [
    {name = "fovegage",email = "fovegage@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-socketio (>=5.13.0,<6.0.0)",
    "cachetools (>=6.1.0,<7.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "tldextract (>=5.3.0,<6.0.0)",
    "curl-cffi (>=0.12.0,<0.13.0)",
    "drissionpage (>=4.1.0.18,<5.0.0.0)",
    "aiohttp (>=3.12.14,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "mirrors"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"
